Refactor package dir
* Rename connector dir to lib

* move connector.js to server dir
