+++ lib/facet-handler.js
