+++ lib/middleware-handler.js
