+++ lib/template-handler.js
