+++ lib/workspace-handler.js
